+++ types/v1/check.go
@@ -1,6 +1,8 @@
 package v1
 
 // CheckResult contains the 1.x compatible check result payload
+// This is a subset of 1.x attributes, see:
+// https://docs.sensu.io/sensu-core/1.6/reference/checks/#check-attributes
 type CheckResult struct {
 	Client      string   `json:"client"`
 	Status      uint32   `json:"status"`
@@ -12,4 +14,10 @@ type CheckResult struct {
 	Executed    int64    `json:"executed"`
 	Duration    float64  `json:"duration"`
 	Output      string   `json:"output"`
+
+	// Handler is deprecated but still supported
+	Handler string `json:"handler"`
+
+	// Handlers supercedes Handler
+	Handlers []string `json:"handlers"`
 }